intent-market-model/pyproject.toml
[project]
name = "intent-market-model"
version = "0.1.0"
description = "Intent-Level Market Model MVP"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "QuantAI" }]

dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "sqlalchemy>=2.0.0",
  "psycopg[binary]>=3.1.0",
  "pgvector>=0.2.5",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.1",
  "typer>=0.12.0",
  "numpy>=1.26.0",
  "scikit-learn>=1.4.0",
  "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
test = [
  "pytest>=8.0.0",
]

[project.scripts]
intent-cli = "app.cli:app"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
